+++ lib/openapi_flutter_gen.dart
@@ -2,5 +2,6 @@ library openapi_flutter_gen;
 
 export 'src/parser/loader.dart';
 export 'src/parser/openapi_parser.dart';
+export 'src/parser/swagger_normalizer.dart';
 export 'src/generator/generator.dart';
 export 'src/ir/ir.dart';